reorganize entry test file
