+++ karma.conf.js
@@ -18,7 +18,28 @@ module.exports = function(config) {
       "es6-shim"
     ],
     files: [
-      "./test/entry.test.ts",
+      // "./test/**/**/commandExecutor.test.ts",
+      // "./test/**/**/commandMapper.test.ts",
+      // "./test/**/**/commandMapping.test.ts",
+      // "./test/**/**/commandMappingList.test.ts",
+      // "./test/**/**/commandPayload.test.ts",
+      // "./test/**/**/commandTriggerMap.test.ts",
+      // "./test/**/**/directCommandMap.test.ts",
+      // "./test/**/**/injectableLoggerExtension.test.ts",
+      // "./test/**/**/applyHooks.test.ts",
+      // "./test/**/**/configManager.test.ts",
+      // "./test/**/**/context.test.ts",
+      // "./test/**/**/extensionInstaller.test.ts",
+      // "./test/**/**/guardsApprove.test.ts",
+      // "./test/**/**/lifecycle.test.ts",
+      // "./test/**/**/lifecycleTransition.test.ts",
+      // "./test/**/**/logger.test.ts",
+      // "./test/**/**/logManager.test.ts",
+      // "./test/**/**/messageDispatcher.test.ts",
+      // "./test/**/**/objectProcessor.test.ts",
+      // "./test/**/**/pin.test.ts",
+      // "./test/**/**/robotlegsInjector.test.ts",
+      // "./test/**/**/safelyCallBack.test.ts",
       "./test/**/**/**.test.ts",
       {
         pattern: '**/*.map',
