+++ test/entry.ts
@@ -1,3 +1,5 @@
+/// <reference types="mocha" />
+
 import "reflect-metadata";
 import "bluebird/js/browser/bluebird";
 import "es6-symbol/implement";
