+++ test/robotlegs/bender/extensions/commandCenter/impl/commandTriggerMap.test.ts
@@ -5,6 +5,8 @@
 //  in accordance with the terms of the license agreement accompanying it.
 // ------------------------------------------------------------------------------
 
+import "../../../../../entry.ts";
+
 import { assert } from "chai";
 
 import sinon = require("sinon");
