+++ test/robotlegs/bender/framework/impl/configManager.test.ts
@@ -5,6 +5,8 @@
 //  in accordance with the terms of the license agreement accompanying it.
 // ------------------------------------------------------------------------------
 
+import "../../../../entry.ts";
+
 import { assert } from "chai";
 
 import { IConfig } from "../../../../../src/robotlegs/bender/framework/api/IConfig";
