+++ tsconfig.json
@@ -3,6 +3,7 @@
     "outDir": "lib",
     "target": "es5",
     "lib": ["es6", "dom"],
+    "types": ["reflect-metadata"],
     "module": "commonjs",
     "moduleResolution": "node",
     "declaration": true,
